requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dacite==1.8.1
Deprecated==1.2.14
discord==2.3.2
discord.py==2.3.2
fastapi==0.103.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
jaconv==0.3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
mutagen==1.47.0
platformdirs==3.11.0
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pykakasi==2.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytube==15.0.0
rapidfuzz==2.15.1
redis==5.0.1
requests==2.31.0
rich==13.6.0
six==1.16.0
sniffio==1.3.0
soundcloud-v2==1.3.1
soupsieve==2.5
spotdl==4.2.1
spotipy==2.23.0
starlette==0.27.0
syncedlyrics==0.5.0
text-unidecode==1.3
typing_extensions==4.8.0
Unidecode==1.3.7
urllib3==2.0.6
uvicorn==0.23.2
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
yt-dlp==2023.10.13
ytmusicapi==1.3.0
